Fix to make it one loop
I believe that Python is exhausting the file on one foreach loop, even though we are breaking early, so the second loop can't continue where we left off. Instead we will have to have a branch statement detect each iteration whether or not we found the names already, and if so start searching for the dates.